# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Mini ERP Backend API doc",
        "title": "Mini ERP Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/products/{id}": {
            "get": {
                "description": "Get product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get Product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/query.ProductResult"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a product by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mini-erp-backend_api_service_product_command.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mini-erp-backend_api_service_product_command.UpdateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid input",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found: Product does not exist",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: Product already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a product by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mini-erp-backend_api_service_product_command.DeleteByIdResult"
                        }
                    }
                }
            }
        },
        "/products/{id}/stock-summary": {
            "get": {
                "description": "Get product stock summary by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get Product Stock Summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/query.ProductStockSummaryResult"
                        }
                    }
                }
            }
        },
...
        "/purchase-orders/{id}/status": {
            "put": {
                "description": "Update the status of a purchase order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PurchaseOrder"
                ],
                "summary": "Update purchase order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase Order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/command.UpdatePOStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/reports/purchase-summary/export": {
            "get": {
                "description": "Export purchase order summary by month to Excel file",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Export purchase report to Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month (MM-YYYY)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reports/stock-movements": {
            "get": {
                "description": "Get stock movements within a date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get stock movements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (DD-MM-YYYY)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To date (DD-MM-YYYY)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/query.StockMovementsResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/reports/stock-summary/export": {
            "get": {
                "description": "Export current stock summary to CSV file",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Export stock summary to CSV",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "command.UpdatePOStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/model.PurchaseOrderStatus"
                }
            }
        },
...
        "mini-erp-backend_api_service_product_command.DeleteByIdResult": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "mini-erp-backend_api_service_product_command.UpdateRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "cost_price": {
                    "type": "number"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "selling_price": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "mini-erp-backend_api_service_product_command.UpdateResult": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/model.Product"
                }
            }
        },
        "mini-erp-backend_api_service_product_query.StockSummary": {
            "type": "object",
            "properties": {
                "current_stock": {
                    "type": "integer"
                },
                "total_adjust": {
                    "type": "integer"
                },
                "total_in": {
                    "type": "integer"
                },
                "total_out": {
                    "type": "integer"
                }
            }
        },
        "model.Category": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "category_id": {
                    "type": "string"
                },
                "cost_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "min_stock": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "product_code": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "selling_price": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.PurchaseOrderStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "CONFIRMED",
                "RECEIVED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "Draft",
                "Confirmed",
                "Received",
                "Cancelled"
            ]
        },
...
        "query.ProductResult": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/model.Product"
                }
            }
        },
        "query.ProductStockSummaryResult": {
            "type": "object",
            "properties": {
                "is_low_stock": {
                    "type": "boolean"
                },
                "min_stock": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "stock_summary": {
                    "$ref": "#/definitions/mini-erp-backend_api_service_product_query.StockSummary"
                }
            }
        },
...
        "query.StockMovementsResult": {
            "type": "object",
            "properties": {
                "movements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.StockMovementResult"
                    }
                }
            }
        },
...
        "repository.StockMovementResult": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "product_code": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "stock_transaction_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Example of \"Value\": Bearer \u003cyour_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}